Раздробил main.js на модули, сделав этот файл входной точкой. Добавил к подключению этого файла в index.html тип модуля. Удалил файл functions.js.
